Project_3/requirements.txt
# Project 3: Marketing Ops Automation
# OpenAI Agents SDK + Notion CRM Integration

# Core dependencies
openai>=1.50.0
openai-agents>=0.6.0  # OpenAI Agents SDK for multi-agent workflows (latest: 0.6.4)
python-dotenv>=1.0.0
pydantic>=2.5.0
notion-client>=2.0.0
redis>=5.0.0  # Session storage for conversational Slack

# Automated triggers
watchdog>=3.0.0
flask>=2.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.10.0

# Optional: PDF export
# reportlab>=4.0.0

# Optional: Zapier MCP integration
# httpx>=0.27.0
